requirements.txt
Django==1.9.2
django-debug-toolbar==1.4
djangorestframework==3.3.2
sqlparse==0.1.18
wheel==0.24.0
